CSRMatrix 필드 수정
row_ptr, diag_ptr, col_indices 필드 Vec<u32> 형으로 변환

+++ src/linear_algebra/msolver/gauss_seidel.rs
@@ -97,23 +97,23 @@ impl MSolver for GaussSeidel {
         while *residual > tol && *iter < iter_max {
             unsafe {
                 for i in 0..n {
-                    let start = *ia.get_unchecked(i);
-                    let end = *ia.get_unchecked(i + 1);
-                    let diag_idx = *da.get_unchecked(i);
+                    let start = *ia.get_unchecked(i) as usize;
+                    let end = *ia.get_unchecked(i + 1) as usize;
+                    let diag_idx = *da.get_unchecked(i) as usize;
 
                     let mut sum = *b.get_unchecked(i);
 
                     let aa_slice = aa.get_unchecked(start..diag_idx);
                     let ja_slice = ja.get_unchecked(start..diag_idx);
                     for (&a_val, &col_idx) in aa_slice.iter().zip(ja_slice.iter()) {
-                        sum -= a_val * x.get_unchecked(col_idx);
+                        sum -= a_val * x.get_unchecked(col_idx as usize);
                         // sum = (-a_val).mul_add(*x.get_unchecked(col_idx), sum);
                     }
 
                     let aa_slice = aa.get_unchecked(diag_idx + 1..end);
                     let ja_slice = ja.get_unchecked(diag_idx + 1..end);
                     for (&a_val, &col_idx) in aa_slice.iter().zip(ja_slice.iter()) {
-                        sum -= a_val * x.get_unchecked(col_idx);
+                        sum -= a_val * x.get_unchecked(col_idx as usize);
                         // sum = (-a_val).mul_add(*x.get_unchecked(col_idx), sum);
                     }
